proj command copy feature WIP
